+++ src/main/java/net/frozenblock/wilderwild/networking/WilderClientNetworking.java
@@ -7,7 +7,6 @@ import net.frozenblock.wilderwild.networking.packet.WilderFloatingSculkBubblePar
 import net.frozenblock.wilderwild.networking.packet.WilderJellyfishStingPacket;
 import net.frozenblock.wilderwild.networking.packet.WilderLightningStrikePacket;
 import net.frozenblock.wilderwild.networking.packet.WilderSensorHiccupPacket;
-import net.frozenblock.wilderwild.networking.packet.WilderTermiteParticlePacket;
 
 @Environment(EnvType.CLIENT)
 public class WilderClientNetworking {
@@ -15,7 +14,6 @@ public class WilderClientNetworking {
 	public static void registerPacketReceivers() {
 		WilderFloatingSculkBubbleParticlePacket.receive();
 		WilderControlledSeedParticlePacket.receive();
-		WilderTermiteParticlePacket.receive();
 		WilderSensorHiccupPacket.receive();
 		WilderJellyfishStingPacket.receive();
 		WilderLightningStrikePacket.receive();